Fix 'division by zero' error + new describe field
When passing the testing dataset as parameter for `describe.py`, the
`Hogwarts House` column contain null values leading to a division by
zero. To counter that, a check on the len of the column is added.

The field `Nulls` has also been added and represents the number of null
values in the colums